refactor(web): remove filter barrel imports
