--- apps/web/src/components/filters/index.ts
@@ -1,4 +1,0 @@
-export { LevelRangeFilter } from "./level-range-filter";
-export { PeriodSelector } from "./period-selector";
-export { CharacterSelector } from "./character-selector";
-export { WarriorSearchFilter } from "./warrior-search-filter";
