+++ apps/web/src/features/user/battle-panel/battle-panel-battles-list/components/filters-sidebar.tsx
@@ -20,7 +20,7 @@ import {
 } from "lucide-react";
 import { cn } from "@lootlog/ui/lib/utils";
 import { FilterPopover } from "@lootlog/ui/components/filter-popover";
-import { LevelRangeFilter } from "@/components/filters";
+import { LevelRangeFilter } from "@/components/filters/level-range-filter";
 import { useBattlesControllerGetUserWorlds } from "@/lib/api/generated/battlelog/battles/battles";
 import { capitalizeFirstLetter } from "@/utils/capitalize-first-letter";
 import {
