+++ apps/web/src/features/user/battle-panel/battle-panel-statistics/player-vs-player-full-page.tsx
@@ -1,5 +1,6 @@
 import { PlayerTile } from "@/components/battle";
-import { LevelRangeFilter, PeriodSelector } from "@/components/filters";
+import { LevelRangeFilter } from "@/components/filters/level-range-filter";
+import { PeriodSelector } from "@/components/filters/period-selector";
 import { TanStackTableBody } from "@/components/ui/tanstack-table-body";
 import { TanStackTableHeader } from "@/components/ui/tanstack-table-header";
 import { TableRowsSkeleton } from "@/components/ui/table-rows-skeleton";
